+++ mpich2/src/binding/f77/attr_deletef.c
@@ -185,5 +185,5 @@ extern FORT_DLL_SPEC void FORT_CALL pmpi_attr_delete( MPI_Fint *, MPI_Fint *, MP
 /* Prototypes for the Fortran interfaces */
 #include "fproto.h"
 FORT_DLL_SPEC void FORT_CALL mpi_attr_delete_ ( MPI_Fint *v1, MPI_Fint *v2, MPI_Fint *ierr ){
-    *ierr = MPI_Attr_delete( (MPI_Comm)(*v1), *v2 );
+    *ierr = MPI_Attr_delete( (MPI_Comm)(*v1), (int)*v2 );
 }
