+++ mpich2/src/binding/f77/comm_acceptf.c
@@ -195,6 +195,6 @@ FORT_DLL_SPEC void FORT_CALL mpi_comm_accept_ ( char *v1 FORT_MIXED_LEN(d1), MPI
         for (li=0; li<(p-v1); li++) { p1[li] = v1[li]; }
         p1[li] = 0; 
     }
-    *ierr = MPI_Comm_accept( p1, (MPI_Info)(*v2), *v3, (MPI_Comm)(*v4), (MPI_Comm *)(v5) );
+    *ierr = MPI_Comm_accept( p1, (MPI_Info)(*v2), (int)*v3, (MPI_Comm)(*v4), (MPI_Comm *)(v5) );
     MPIU_Free( p1 );
 }
