+++ mpich2/src/binding/f77/commcallerrf.c
@@ -185,5 +185,5 @@ extern FORT_DLL_SPEC void FORT_CALL pmpi_comm_call_errhandler( MPI_Fint *, MPI_F
 /* Prototypes for the Fortran interfaces */
 #include "fproto.h"
 FORT_DLL_SPEC void FORT_CALL mpi_comm_call_errhandler_ ( MPI_Fint *v1, MPI_Fint *v2, MPI_Fint *ierr ){
-    *ierr = MPI_Comm_call_errhandler( (MPI_Comm)(*v1), *v2 );
+    *ierr = MPI_Comm_call_errhandler( (MPI_Comm)(*v1), (int)*v2 );
 }
