+++ mpich2/src/binding/f77/dist_graph_neighborsf.c
@@ -189,5 +189,5 @@ FORT_DLL_SPEC void FORT_CALL mpi_dist_graph_neighbors_ ( MPI_Fint *v1, MPI_Fint 
 #ifndef HAVE_MPI_F_INIT_WORKS_WITH_C
     if (MPIR_F_NeedInit){ mpirinitf_(); MPIR_F_NeedInit = 0; }
 #endif
-    *ierr = MPI_Dist_graph_neighbors( (MPI_Comm)(*v1), *v2, v3, v4, *v5, v6, v7 );
+    *ierr = MPI_Dist_graph_neighbors( (MPI_Comm)(*v1), (int)*v2, v3, v4, (int)*v5, v6, v7 );
 }
