+++ mpich2/src/binding/f77/file_read_all_beginf.c
@@ -186,7 +186,7 @@ extern FORT_DLL_SPEC void FORT_CALL pmpi_file_read_all_begin( MPI_Fint *, void*,
 #include "fproto.h"
 FORT_DLL_SPEC void FORT_CALL mpi_file_read_all_begin_ ( MPI_Fint *v1, void*v2, MPI_Fint *v3, MPI_Fint *v4, MPI_Fint *ierr ){
 #ifdef MPI_MODE_RDONLY
-    *ierr = MPI_File_read_all_begin( MPI_File_f2c(*v1), v2, *v3, (MPI_Datatype)(*v4) );
+    *ierr = MPI_File_read_all_begin( MPI_File_f2c(*v1), v2, (int)*v3, (MPI_Datatype)(*v4) );
 #else
 *ierr = MPI_ERR_INTERN;
 #endif
