+++ mpich2/src/binding/f77/grfnbcountf.c
@@ -185,5 +185,5 @@ extern FORT_DLL_SPEC void FORT_CALL pmpi_graph_neighbors_count( MPI_Fint *, MPI_
 /* Prototypes for the Fortran interfaces */
 #include "fproto.h"
 FORT_DLL_SPEC void FORT_CALL mpi_graph_neighbors_count_ ( MPI_Fint *v1, MPI_Fint *v2, MPI_Fint *v3, MPI_Fint *ierr ){
-    *ierr = MPI_Graph_neighbors_count( (MPI_Comm)(*v1), *v2, v3 );
+    *ierr = MPI_Graph_neighbors_count( (MPI_Comm)(*v1), (int)*v2, v3 );
 }
