+++ mpich2/src/binding/f77/sndrcvrplf.c
@@ -191,5 +191,5 @@ FORT_DLL_SPEC void FORT_CALL mpi_sendrecv_replace_ ( void*v1, MPI_Fint *v2, MPI_
 #endif
 
     if (v9 == MPI_F_STATUS_IGNORE) { v9 = (MPI_Fint*)MPI_STATUS_IGNORE; }
-    *ierr = MPI_Sendrecv_replace( v1, *v2, (MPI_Datatype)(*v3), *v4, *v5, *v6, *v7, (MPI_Comm)(*v8), (MPI_Status *)v9 );
+    *ierr = MPI_Sendrecv_replace( v1, (int)*v2, (MPI_Datatype)(*v3), (int)*v4, (int)*v5, (int)*v6, (int)*v7, (MPI_Comm)(*v8), (MPI_Status *)v9 );
 }
