+++ mpich2/src/binding/f77/typedelattrf.c
@@ -185,5 +185,5 @@ extern FORT_DLL_SPEC void FORT_CALL pmpi_type_delete_attr( MPI_Fint *, MPI_Fint 
 /* Prototypes for the Fortran interfaces */
 #include "fproto.h"
 FORT_DLL_SPEC void FORT_CALL mpi_type_delete_attr_ ( MPI_Fint *v1, MPI_Fint *v2, MPI_Fint *ierr ){
-    *ierr = MPI_Type_delete_attr( (MPI_Datatype)(*v1), *v2 );
+    *ierr = MPI_Type_delete_attr( (MPI_Datatype)(*v1), (int)*v2 );
 }
